+++ oak-store-document/src/main/java/org/apache/jackrabbit/oak/plugins/document/rdb/RDBDocumentStore.java
@@ -2104,7 +2104,7 @@ public class RDBDocumentStore implements DocumentStore {
             if (!success && shouldRetry) {
                 data = ser.asString(document, tmd.getColumnOnlyProperties());
                 Object m = document.get(MODIFIED);
-                long modified = (m instanceof Long) ? ((Long)m).longValue() : 0;
+                long modified = (m instanceof Long) ? (Long) m : 0;
                 success = db.update(connection, tmd, document.getId(), modified, hasBinary, deletedOnce, modcount, cmodcount,
                         oldmodcount, data);
                 connection.commit();
@@ -2319,7 +2319,7 @@ public class RDBDocumentStore implements DocumentStore {
 
     private static long modifiedOf(@NotNull Document doc) {
         Object l = doc.get(NodeDocument.MODIFIED_IN_SECS);
-        return (l instanceof Long) ? ((Long)l).longValue() : -1;
+        return (l instanceof Long) ? (Long) l : -1;
     }
 
     @NotNull
