+++ oak-store-document/src/test/java/org/apache/jackrabbit/oak/plugins/document/BulkCreateOrUpdateTest.java
@@ -336,7 +336,7 @@ public class BulkCreateOrUpdateTest extends AbstractDocumentStoreTest {
             if (prevModCount != null) {
                 assertNotNull(modCount);
                 assertTrue("_modCount, when present, must be increasing, but changed from " + prevModCount + " to " + modCount,
-                        prevModCount.longValue() < modCount.longValue());
+                        prevModCount < modCount);
             }
             prevModCount = modCount;
             assertEquals("The old value is not correct", (long) (i - 1), docs.get(i).get("update_id"));
